pyproject.toml
[tool.poetry]
name = 'blogproject'
version = '0.1.0'
description = ""
authors = ['yan <korotkevich3007@gmail.com>']
readme = 'README.md'

[tool.poetry.dependencies]
python = '^3.12'
requests = "^2.32.3"
aiogram = "^3.15.0"
python-dotenv = "^1.0.1"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
pillow = "^11.1.0"

[tool.black]
line-length = 130
target-version = ['py37']

[tool.isort]
profile = "black"